+++ packages/extension-web-ui/src/hooks/ledger/useLedger.ts
@@ -5,6 +5,7 @@ import { _ChainInfo } from '@subwallet/chain-list/types';
 import { LedgerNetwork } from '@subwallet/extension-base/background/KoniTypes';
 import { _isChainEvmCompatible } from '@subwallet/extension-base/services/chain-service/utils';
 import { EVMLedger, SubstrateLedger } from '@subwallet/extension-web-ui/connector';
+import { isLedgerCapable } from '@subwallet/extension-web-ui/constants';
 import { useSelector } from '@subwallet/extension-web-ui/hooks';
 import useGetSupportedLedger from '@subwallet/extension-web-ui/hooks/ledger/useGetSupportedLedger';
 import { Ledger } from '@subwallet/extension-web-ui/types';
@@ -34,8 +35,6 @@ interface Result extends StateBase {
   signMessage: Ledger['signMessage'];
 }
 
-const isLedgerCapable = !!(window as unknown as { USB?: unknown }).USB;
-
 const baseState: StateBase = {
   isLedgerCapable,
   isLedgerEnabled: isLedgerCapable
@@ -47,13 +46,10 @@ const getNetwork = (ledgerChains: LedgerNetwork[], slug: string, isEthereumNetwo
   return ledgerChains.find((network) => network.slug === slug || (network.isEthereum && isEthereumNetwork));
 };
 
-const retrieveLedger = (slug: string, ledgerChains: LedgerNetwork[], chainInfoMap: Record<string, _ChainInfo>): Ledger => {
+const retrieveLedger = (slug: string, ledgerChains: LedgerNetwork[], isEthereumNetwork: boolean): Ledger => {
   const { isLedgerCapable } = baseState;
 
   assert(isLedgerCapable, 'Incompatible browser, only Chrome is supported');
-
-  const chainInfo = chainInfoMap[slug];
-  const isEthereumNetwork = _isChainEvmCompatible(chainInfo);
 
   const def = getNetwork(ledgerChains, slug, isEthereumNetwork);
 
@@ -71,6 +67,16 @@ export function useLedger (slug?: string, active = true): Result {
 
   const ledgerChains = useGetSupportedLedger();
   const { chainInfoMap } = useSelector((state) => state.chainStore);
+
+  const isEvmNetwork = useMemo(() => {
+    if (!slug) {
+      return false;
+    }
+
+    const chainInfo = chainInfoMap[slug];
+
+    return _isChainEvmCompatible(chainInfo);
+  }, [chainInfoMap, slug]);
 
   const timeOutRef = useRef<NodeJS.Timer>();
   const destroyRef = useRef<VoidFunction>();
@@ -98,14 +104,14 @@ export function useLedger (slug?: string, active = true): Result {
       }
 
       try {
-        return retrieveLedger(slug, ledgerChains, chainInfoMap);
+        return retrieveLedger(slug, ledgerChains, isEvmNetwork);
       } catch (error) {
         setError((error as Error).message);
       }
     }
 
     return null;
-  }, [refreshLock, slug, active, ledgerChains, chainInfoMap]);
+  }, [refreshLock, slug, active, ledgerChains, isEvmNetwork]);
 
   const appName = useMemo(() => {
     const unknownNetwork = 'unknown network';